minimized mapping exports
Renamed functions in handlers files to coincide with definitions in
Manifest
Eported all handlers from main mapping file
Removed all handlers named imports and all wrappers code
